Type:chore
Description:警告处理

RootCause:Null

Solution:

+++ dialogs/dtaskdialog.cpp
@@ -97,7 +97,7 @@ DFileCopyMoveJob::Action ErrorHandle::handleError(DFileCopyMoveJob *job, DFileCo
 static QString formatTime(int second)
 {
     quint8 s = second % 60;
-    quint8 m = second / 60;
+    quint8 m = static_cast<quint8>(second / 60);
     quint8 h = m / 60;
     quint8 d = h / 24;
